requirements.txt
fastapi                                                                                                                                                              
pydantic                                                                                                                                                             
requests                                                                                                                                                             
fastapi-cache2                                                                                                                                                       
redis                                                                                                                                                                
nltk                                                                                                                                                                 
joblib                                                                                                                                                               
pandas          
